requirements.txt
fastapi==0.129.1
uvicorn==0.41.0
jinja2==3.1.6
python-multipart==0.0.22
pandas==2.3.3
python-dotenv==1.2.2
neo4j==6.1.0
weaviate-client==4.20.2
sentence-transformers==5.2.3
transformers==5.2.0
torch==2.9.1
groq==1.0.0
rouge==1.0.1
bert-score==0.3.13
scikit-learn==1.7.2
